Try to fix Visual Studio compiler warning
I'm not completely happy with this - I'm happy it's correct, but it
feels a bit "lazy" to do the cast here rather than propagate the change
outward, but so much other stuff this ties in with uses int that this
seems the least disruptive way to fix it.

+++ src/rivers_rainfall.h
@@ -773,7 +773,7 @@ struct Lake {
 	/** Returns the number of tiles of the lake.
 	 *  @return the number of tiles of the lake.
 	 */
-	inline int GetNumberOfLakeTiles() { return this->lake_tiles.size(); }
+	inline int GetNumberOfLakeTiles() { return static_cast<int>(this->lake_tiles.size()); }
 
 	/** Returns the begin() const iterator for the lake tiles.
 	 *  @return the begin() const iterator for the lake tiles.
